fixed '+=' list build syntax
Changed syntax
forEach list_item
    list.innerHTML += list_item
To
forEach list_item
    list_string += list_item
list.innerHTML = list_string
So the inner HTML is set only once. This seems to have fixed the bottleneck that locked up the species search. This means the RxJS is not needed, and RxJS would not have fixed the problem anyway since the bottleneck was not in generating the arrays but putting them into the lists.
Also fixed bug that was putting site info cards inside earlier ones